Add dataset argument to pixelcnn_prior
